requirements.txt
aiogram>=3.20.0
fastapi>=0.111.0
uvicorn>=0.30.0
motor>=3.5.1
pydantic>=2.11.7
python-dotenv>=1.0.1
payos>=1.1.0
qrcode>=7.4.2
Pillow>=11.0.0
loguru
